fix: rename footer imports
