+++ src/components/pages/HomePage.tsx
@@ -3,7 +3,7 @@ import AccountSummary from '../common/AccountSummary';
 import { Container } from '@mui/material';
 import SpendFrequencyChart from '../common/SpendFrequencyChart';
 import RecentTransactions from '../common/RecentTransactions';
-import FooterWithFab from '../common/HomeFooter';
+import FooterWithFab from '../common/FooterWithFab/FooterWithFab';
 
 const HomePage: React.FC = () => {
   return (
